convert ThroughputPerformance to FC
